refactor: split fund results presentation from route shell
The route shell should compose server-backed results, while section cards own their rendering details. This keeps evidence and section copy reviewable without changing contracts.

Constraint: Preserve rendered copy, evidence test ids, scenario comparison rendering, and lifecycle card semantics.

Confidence: medium

Scope-risk: moderate

Tested: npx vitest client fund-model-results page plus evidence and fund-results component tests

Not-tested: Full client suite
