+++ client/src/pages/fund-model-results.tsx
@@ -11,22 +11,9 @@
  */
 
 import React, { useRef, useState, useEffect, useCallback } from 'react';
-import { Link, useRoute } from 'wouter';
-import { AlertCircle, ChevronDown, ChevronRight, History, RefreshCw } from 'lucide-react';
-import { Alert, AlertTitle, AlertDescription } from '@/components/ui/alert';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
-import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible';
-import {
-  CrossSetScenarioComparisonTable,
-  isComparableEconomicsComparison,
-  ScenarioComparisonTable,
-  ScenarioSetsSummary,
-} from '@/components/fund-results';
-import { EvidenceHeader, type EvidenceHeaderLifecycle } from '@/components/results/EvidenceHeader';
+import { useRoute } from 'wouter';
 import { QuarterlyReviewTrace } from '@/features/analytics-parity/QuarterlyReviewTrace';
 import { apiRequest, queryClient } from '@/lib/queryClient';
-import { cn } from '@/lib/utils';
 import type {
   FundResultsReadV1,
   ScorecardPayload,
@@ -38,40 +25,29 @@ import {
   FundScenarioComparisonV1Schema,
   type FundScenarioComparisonV1,
 } from '@shared/contracts/fund-scenario-comparison-v1.contract';
-import type { FundStateReadV1 } from '@shared/contracts/fund-state-read-v1.contract';
 import type { FundLifecycleHistoryV1 } from '@shared/contracts/fund-lifecycle-history-v1.contract';
-import type {
-  FundResultsComparisonV1,
-  PublishedVersionSummary,
-} from '@shared/contracts/fund-results-comparison-v1.contract';
+import type { FundResultsComparisonV1 } from '@shared/contracts/fund-results-comparison-v1.contract';
 import {
   configBackedEvidence,
   evidenceFromLifecycle,
-  getLifecycleDiagnostic,
   mixedScorecardEvidence,
-  reasonCopyFor,
   sectionBackedEvidence,
-  sectionEvidence,
 } from './fund-model-results/evidence';
 import { FadeInSection } from './fund-model-results/FadeInSection';
-import { ErrorState, LatestErrorState, LoadingState } from './fund-model-results/states';
 import {
-  capitalize,
-  diagnosticAlertClasses,
-  formatCompactMoney,
-  formatComparisonDelta,
-  formatComparisonMetricValue,
-  formatDateOrFallback,
-  formatDriftCapabilityReason,
-  formatHistoryRunStatus,
-  formatLifecycleStatus,
-  formatMultiple,
-  formatNullablePercent,
-  hasStaleEvidence,
-  historyBadgeClasses,
-  percent,
-  percentPoints,
-} from './fund-model-results/formatters';
+  ConfigDiffBanner,
+  LifecycleStatusCard,
+  PublishComparisonCard,
+  PublishHistoryCard,
+} from './fund-model-results/lifecycle-cards';
+import {
+  EconomicsResultsCard,
+  OverviewCard,
+  WaterfallSetupCard,
+} from './fund-model-results/result-section-cards';
+import { ScenarioAnalysisCard } from './fund-model-results/scenario-section';
+import { SectionRenderer } from './fund-model-results/SectionRenderer';
+import { ErrorState, LatestErrorState, LoadingState } from './fund-model-results/states';
 import type {
   FetchOptions,
   FetchState,
@@ -574,893 +550,6 @@ function useRecalculatePublished(fundId: string | null, onSuccess: () => void) {
     recalculate,
     clearError: () => setState({ kind: 'idle' }),
   };
-}
-
-// ============================================================================
-// OVERVIEW (SCORECARD) CARD
-// ============================================================================
-
-function OverviewCard({ payload }: { payload: ScorecardPayload }) {
-  return (
-    <div className="grid grid-cols-2 md:grid-cols-3 gap-4">
-      <FactTile label="Fund Size" value={`$${(payload.fundSize.value / 1_000_000).toFixed(0)}M`} />
-      {payload.vintageYear && (
-        <FactTile label="Vintage Year" value={String(payload.vintageYear.value)} />
-      )}
-      {payload.reserveRatio && (
-        <FactTile
-          label="Reserve Ratio"
-          value={`${(payload.reserveRatio.value * 100).toFixed(1)}%`}
-        />
-      )}
-      {payload.avgConfidence && (
-        <FactTile
-          label="Avg Confidence"
-          value={`${(payload.avgConfidence.value * 100).toFixed(0)}%`}
-        />
-      )}
-      {payload.yearsToFullDeploy && (
-        <FactTile label="Years to Full Deploy" value={`${payload.yearsToFullDeploy.value} yrs`} />
-      )}
-      {payload.lastCalculatedAt && (
-        <FactTile
-          label="Last Calculated"
-          value={new Date(payload.lastCalculatedAt.value).toLocaleDateString()}
-        />
-      )}
-    </div>
-  );
-}
-
-function WaterfallSetupCard({ payload }: { payload: WaterfallSetupSection }) {
-  return (
-    <div className="space-y-5">
-      <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-        <FactTile label="Structure" value={capitalize(payload.type)} />
-        <FactTile label="Tiers" value={String(payload.tierCount)} />
-        <FactTile
-          label="Recycling"
-          value={
-            payload.recyclingEnabled == null
-              ? 'Not set'
-              : payload.recyclingEnabled
-                ? 'Enabled'
-                : 'Disabled'
-          }
-        />
-        <FactTile
-          label="Recycling Type"
-          value={payload.recyclingType ? capitalize(payload.recyclingType) : 'Not set'}
-        />
-      </div>
-
-      <div className="space-y-3">
-        {payload.tiers.map((tier, index) => (
-          <div key={`${tier.name}-${index}`} className="rounded-md border border-beige-200 p-4">
-            <div className="flex items-center justify-between gap-4">
-              <div>
-                <p className="font-medium text-charcoal">{tier.name}</p>
-                <p className="text-sm text-charcoal-500 font-poppins">
-                  GP {tier.gpSplit}% / LP {tier.lpSplit}%
-                </p>
-              </div>
-              {tier.condition && tier.condition !== 'none' && tier.conditionValue != null && (
-                <p className="text-sm text-charcoal-500 font-poppins">
-                  {tier.condition.toUpperCase()} hurdle {tier.conditionValue}
-                </p>
-              )}
-            </div>
-            <div className="mt-3 grid grid-cols-2 gap-3 md:grid-cols-4">
-              <FactTile
-                label="Preferred Return"
-                value={tier.preferredReturn != null ? percent(tier.preferredReturn) : 'Not set'}
-              />
-              <FactTile
-                label="Catch-up"
-                value={tier.catchUp != null ? percent(tier.catchUp) : 'Not set'}
-              />
-              <FactTile
-                label="Recycling Cap"
-                value={
-                  payload.recyclingCap != null ? percentPoints(payload.recyclingCap) : 'Not set'
-                }
-              />
-              <FactTile
-                label="Future Recycling"
-                value={
-                  payload.allowFutureRecycling == null
-                    ? 'Not set'
-                    : payload.allowFutureRecycling
-                      ? 'Allowed'
-                      : 'Not allowed'
-                }
-              />
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function ScenarioComparisonPanel({ state }: { state: ScenarioComparisonState }) {
-  if (state.kind === 'idle') return null;
-
-  const comparable = state.comparisons.filter(isComparableEconomicsComparison);
-  const nonComparable = state.comparisons.filter(
-    (comparison) => !isComparableEconomicsComparison(comparison)
-  );
-  const hasContent = state.comparisons.length > 0 || state.failedScenarioSetIds.length > 0;
-
-  return (
-    <div className="space-y-4">
-      {state.kind === 'loading' && !hasContent && (
-        <p className="text-sm text-charcoal-500 font-poppins">Loading scenario comparison...</p>
-      )}
-
-      {state.kind === 'loading' && hasContent && (
-        <p className="text-xs text-charcoal-400 font-poppins">Refreshing scenario comparison...</p>
-      )}
-
-      {state.kind === 'error' && (
-        <Alert className="border-beige-200 bg-beige-50">
-          <AlertCircle className="h-4 w-4 text-charcoal-400" />
-          <AlertTitle>Scenario comparison unavailable</AlertTitle>
-          <AlertDescription className="font-poppins text-charcoal-500">
-            {state.message}
-          </AlertDescription>
-        </Alert>
-      )}
-
-      {comparable.length >= 2 && <CrossSetScenarioComparisonTable comparisons={comparable} />}
-
-      {comparable.length === 1 && comparable[0] && (
-        <ScenarioComparisonTable comparison={comparable[0]} />
-      )}
-
-      {nonComparable.map((comparison) => (
-        <ScenarioComparisonTable
-          key={comparison.scenarioSet.scenarioSetId}
-          comparison={comparison}
-        />
-      ))}
-
-      {state.failedScenarioSetIds.map((scenarioSetId) => (
-        <div
-          key={scenarioSetId}
-          data-testid="scenario-comparison-failed-card"
-          className="rounded-md border border-beige-200 bg-beige-50 p-4"
-        >
-          <p className="text-sm text-charcoal-600 font-poppins">
-            Scenario comparison could not be loaded for this scenario set.
-          </p>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function ScenarioAnalysisCard({
-  fundId,
-  payload,
-  comparisonState,
-}: {
-  fundId: string;
-  payload: ScenariosSectionPayloadV1;
-  comparisonState: ScenarioComparisonState;
-}) {
-  return (
-    <div className="space-y-6">
-      <div className="flex justify-end">
-        <Button asChild variant="outline" size="sm">
-          <Link href={`/fund-model-results/${fundId}/scenarios`}>Open Scenario Workspace</Link>
-        </Button>
-      </div>
-      <ScenarioSetsSummary payload={payload} />
-      <ScenarioComparisonPanel state={comparisonState} />
-    </div>
-  );
-}
-
-function EconomicsResultsCard({ payload }: { payload: EconomicsResultV1 }) {
-  const { summary, annual, checks } = payload;
-
-  return (
-    <div className="space-y-6" data-testid="economics-results-card">
-      <div className="grid grid-cols-2 gap-4 md:grid-cols-4">
-        <FactTile label="Gross IRR" value={formatNullablePercent(summary.grossIrr)} />
-        <FactTile label="Net LP IRR" value={formatNullablePercent(summary.lpNetIrr)} />
-        <FactTile label="Net GP IRR" value={formatNullablePercent(summary.gpNetIrr)} />
-        <FactTile
-          label="Total GP Carry"
-          value={formatCompactMoney(summary.totalGpCarryDistributed)}
-        />
-        <FactTile label="Management Fees" value={formatCompactMoney(summary.totalManagementFees)} />
-        <FactTile label="DPI" value={formatMultiple(summary.finalDpi)} />
-        <FactTile label="TVPI" value={formatMultiple(summary.finalTvpi)} />
-        <FactTile label="Clawback Exposure" value={formatCompactMoney(summary.finalClawbackDue)} />
-      </div>
-
-      <div className="rounded-md border border-beige-200 p-4">
-        <div className="mb-3 flex items-center justify-between gap-4">
-          <h3 className="font-medium text-charcoal">Economics Cashflows</h3>
-          <Badge variant={checks.passed ? 'secondary' : 'destructive'}>
-            {checks.passed ? 'Invariants Passed' : `${checks.errors.length} Invariant Issues`}
-          </Badge>
-        </div>
-        <EconomicsCashflowChart rows={annual} />
-      </div>
-
-      <div className="rounded-md border border-beige-200 p-4">
-        <h3 className="mb-3 font-medium text-charcoal">DPI / RVPI / TVPI</h3>
-        <EconomicsJCurveChart rows={annual} />
-      </div>
-
-      <div className="rounded-md border border-beige-200 p-4">
-        <h3 className="mb-3 font-medium text-charcoal">Waterfall and Carry</h3>
-        <EconomicsCarryTable rows={annual} />
-      </div>
-    </div>
-  );
-}
-
-function EconomicsCashflowChart({ rows }: { rows: EconomicsResultV1['annual'] }) {
-  const maxAbs = Math.max(
-    1,
-    ...rows.flatMap((row) => [
-      row.lpCapitalCalls,
-      row.gpCommitmentCalls,
-      row.lpDistributions,
-      row.gpInvestmentDistributions,
-      row.gpCarryDistributed,
-      row.feesPaidToManager,
-      row.expensesPaid,
-    ])
-  );
-  const series = [
-    { key: 'lpCapitalCalls', label: 'LP Calls', color: 'bg-charcoal-300' },
-    { key: 'gpCommitmentCalls', label: 'GP Calls', color: 'bg-charcoal-500' },
-    { key: 'lpDistributions', label: 'LP Distributions', color: 'bg-presson-positive' },
-    { key: 'gpInvestmentDistributions', label: 'GP Investment', color: 'bg-presson-info' },
-    { key: 'gpCarryDistributed', label: 'GP Carry', color: 'bg-success' },
-    { key: 'feesPaidToManager', label: 'Fees', color: 'bg-presson-warning' },
-    { key: 'expensesPaid', label: 'Expenses', color: 'bg-presson-negative' },
-  ] as const;
-
-  return (
-    <div className="space-y-4">
-      <div className="flex flex-wrap gap-3">
-        {series.map((item) => (
-          <div key={item.key} className="flex items-center gap-2 text-xs text-charcoal-500">
-            <span className={cn('h-2.5 w-2.5 rounded-sm', item.color)} />
-            {item.label}
-          </div>
-        ))}
-      </div>
-      <div className="space-y-3">
-        {rows.map((row) => (
-          <div key={row.year} className="grid grid-cols-[3rem_1fr] items-center gap-3">
-            <div className="text-xs font-medium text-charcoal-500">Y{row.year}</div>
-            <div className="grid h-14 grid-cols-7 items-end gap-1 rounded-md bg-beige-50 px-2 py-1">
-              {series.map((item) => {
-                const value = row[item.key];
-                const heightPct = Math.max(4, (value / maxAbs) * 100);
-                return (
-                  <div
-                    key={item.key}
-                    className={cn('rounded-sm', item.color)}
-                    style={{ height: `${heightPct}%` }}
-                    title={`${item.label}: ${formatCompactMoney(value)}`}
-                  />
-                );
-              })}
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function EconomicsJCurveChart({ rows }: { rows: EconomicsResultV1['annual'] }) {
-  const maxMultiple = Math.max(1, ...rows.map((row) => Math.max(row.dpi, row.rvpi, row.tvpi)));
-  const metrics = [
-    { key: 'dpi', label: 'DPI', color: 'bg-presson-positive' },
-    { key: 'rvpi', label: 'RVPI', color: 'bg-presson-info' },
-    { key: 'tvpi', label: 'TVPI', color: 'bg-charcoal-500' },
-  ] as const;
-
-  return (
-    <div className="space-y-3">
-      {rows.map((row) => (
-        <div key={row.year} className="grid grid-cols-[3rem_1fr] gap-3">
-          <div className="text-xs font-medium text-charcoal-500">Y{row.year}</div>
-          <div className="space-y-1.5">
-            {metrics.map((metric) => (
-              <div key={metric.key} className="grid grid-cols-[3rem_1fr_3rem] items-center gap-2">
-                <span className="text-xs text-charcoal-400">{metric.label}</span>
-                <div className="h-2 overflow-hidden rounded-full bg-beige-100">
-                  <div
-                    className={cn('h-full rounded-full', metric.color)}
-                    style={{ width: `${Math.max(1, (row[metric.key] / maxMultiple) * 100)}%` }}
-                  />
-                </div>
-                <span className="text-right text-xs text-charcoal-500">
-                  {formatMultiple(row[metric.key])}
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function EconomicsCarryTable({ rows }: { rows: EconomicsResultV1['annual'] }) {
-  return (
-    <div className="overflow-x-auto">
-      <table className="min-w-full text-left text-sm">
-        <thead className="border-b border-beige-200 text-xs text-charcoal-400">
-          <tr>
-            <th className="py-2 pr-4 font-medium">Year</th>
-            <th className="py-2 pr-4 font-medium">LP Dist.</th>
-            <th className="py-2 pr-4 font-medium">GP Inv. Dist.</th>
-            <th className="py-2 pr-4 font-medium">GP Carry</th>
-            <th className="py-2 pr-4 font-medium">Escrowed</th>
-            <th className="py-2 pr-4 font-medium">Released</th>
-            <th className="py-2 pr-4 font-medium">Clawback</th>
-            <th className="py-2 pr-4 font-medium">DPI</th>
-            <th className="py-2 pr-4 font-medium">TVPI</th>
-          </tr>
-        </thead>
-        <tbody className="divide-y divide-beige-100">
-          {rows.map((row) => (
-            <tr key={row.year}>
-              <td className="py-2 pr-4 font-medium text-charcoal">Y{row.year}</td>
-              <td className="py-2 pr-4 text-charcoal-500">
-                {formatCompactMoney(row.lpDistributions)}
-              </td>
-              <td className="py-2 pr-4 text-charcoal-500">
-                {formatCompactMoney(row.gpInvestmentDistributions)}
-              </td>
-              <td className="py-2 pr-4 text-charcoal-500">
-                {formatCompactMoney(row.gpCarryDistributed)}
-              </td>
-              <td className="py-2 pr-4 text-charcoal-500">
-                {formatCompactMoney(row.gpCarryEscrowed)}
-              </td>
-              <td className="py-2 pr-4 text-charcoal-500">
-                {formatCompactMoney(row.gpCarryReleasedFromEscrow)}
-              </td>
-              <td className="py-2 pr-4 text-charcoal-500">
-                {formatCompactMoney(row.clawbackPaid)}
-              </td>
-              <td className="py-2 pr-4 text-charcoal-500">{formatMultiple(row.dpi)}</td>
-              <td className="py-2 pr-4 text-charcoal-500">{formatMultiple(row.tvpi)}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-function FactTile({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="bg-beige-50 rounded-md p-3">
-      <p className="text-xs text-charcoal-400 font-poppins">{label}</p>
-      <p className="text-lg font-medium text-charcoal">{value}</p>
-    </div>
-  );
-}
-
-function renderRunSummary(summary: PublishedVersionSummary) {
-  if (!summary.calcRun) {
-    return (
-      <div className="flex items-center gap-2">
-        <Badge variant="outline" className={historyBadgeClasses(null)}>
-          Not started
-        </Badge>
-        <span className="text-sm text-charcoal-500 font-poppins">No calculation run</span>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex flex-wrap items-center gap-2">
-      <Badge variant="outline" className={historyBadgeClasses(summary.calcRun.status)}>
-        {formatLifecycleStatus(summary.calcRun.status)}
-      </Badge>
-      <span className="text-sm text-charcoal-500 font-poppins">Run {summary.calcRun.runId}</span>
-    </div>
-  );
-}
-
-function ConfigDiffBanner({ lifecycle }: { lifecycle: FundStateReadV1 }) {
-  if (!hasStaleEvidence(lifecycle)) return null;
-
-  return (
-    <Alert className="border-warning/30 bg-warning-light">
-      <AlertCircle className="h-4 w-4 text-warning" />
-      <AlertTitle>Results are stale</AlertTitle>
-      <AlertDescription className="font-poppins text-warning-dark">
-        Latest published configuration is v{lifecycle.configState.publishedVersion}, but the current
-        calculation is still on v{lifecycle.calculationState.configVersion}. Recalculate to refresh
-        the published results.
-      </AlertDescription>
-    </Alert>
-  );
-}
-
-function PublishHistoryCard({ historyState }: { historyState: LifecycleHistoryState }) {
-  const [isOpen, setIsOpen] = useState(false);
-  const entryCount =
-    historyState.kind === 'data'
-      ? historyState.history.entries.length
-      : (historyState.history?.entries.length ?? 0);
-
-  return (
-    <div
-      className="bg-white rounded-lg border border-beige-200 p-6 space-y-4"
-      data-testid="publish-history-card"
-    >
-      <Collapsible open={isOpen} onOpenChange={setIsOpen}>
-        <div className="flex flex-col gap-4 md:flex-row md:items-center md:justify-between">
-          <div>
-            <div className="flex items-center gap-2">
-              <History className="h-4 w-4 text-charcoal-500" />
-              <h2 className="text-lg font-medium text-charcoal">Publish History</h2>
-            </div>
-            <p className="mt-1 text-sm text-charcoal-500 font-poppins">
-              {entryCount > 0
-                ? `${entryCount} published version${entryCount === 1 ? '' : 's'} with latest run status`
-                : 'No published versions recorded yet.'}
-            </p>
-          </div>
-
-          <CollapsibleTrigger asChild>
-            <Button variant="outline" size="sm" disabled={entryCount === 0}>
-              {isOpen ? <ChevronDown className="h-4 w-4" /> : <ChevronRight className="h-4 w-4" />}
-              {isOpen ? 'Hide history' : 'Show history'}
-            </Button>
-          </CollapsibleTrigger>
-        </div>
-
-        <CollapsibleContent className="space-y-3 pt-2">
-          {historyState.kind === 'loading' && (
-            <p className="text-sm text-charcoal-500 font-poppins">Loading publish history…</p>
-          )}
-
-          {historyState.kind === 'error' && (
-            <Alert className="border-beige-200">
-              <AlertCircle className="h-4 w-4 text-charcoal-400" />
-              <AlertTitle>Publish history unavailable</AlertTitle>
-              <AlertDescription className="font-poppins text-charcoal-500">
-                {historyState.message}
-              </AlertDescription>
-            </Alert>
-          )}
-
-          {historyState.kind === 'data' &&
-            historyState.history.entries.map((entry) => (
-              <div
-                key={`publish-history-${entry.version}`}
-                className="flex flex-col gap-3 rounded-md border border-beige-200 p-4 md:flex-row md:items-center md:justify-between"
-              >
-                <div>
-                  <p className="font-medium text-charcoal">Version v{entry.version}</p>
-                  <p className="text-sm text-charcoal-500 font-poppins">
-                    Published {formatDateOrFallback(entry.publishedAt)}
-                  </p>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-2 md:justify-end">
-                  <Badge
-                    variant="outline"
-                    className={historyBadgeClasses(entry.calcRun?.status ?? null)}
-                  >
-                    {formatHistoryRunStatus(entry.calcRun?.status ?? null)}
-                  </Badge>
-                  <span className="text-sm text-charcoal-500 font-poppins">
-                    {entry.calcRun?.runId != null
-                      ? `Run ${entry.calcRun.runId}`
-                      : 'No calculation run'}
-                  </span>
-                </div>
-              </div>
-            ))}
-        </CollapsibleContent>
-      </Collapsible>
-    </div>
-  );
-}
-
-function PublishComparisonCard({ comparisonState }: { comparisonState: ResultsComparisonState }) {
-  return (
-    <div
-      className="bg-white rounded-lg border border-beige-200 p-6 space-y-4"
-      data-testid="publish-comparison-card"
-    >
-      <div>
-        <h2 className="text-lg font-medium text-charcoal">Publish Comparison</h2>
-        <p className="mt-1 text-sm text-charcoal-500 font-poppins">
-          Compare the current published version with the immediately previous publication.
-        </p>
-      </div>
-
-      {comparisonState.kind === 'loading' && (
-        <p className="text-sm text-charcoal-500 font-poppins">Loading publish comparison…</p>
-      )}
-
-      {comparisonState.kind === 'error' && (
-        <Alert className="border-beige-200">
-          <AlertCircle className="h-4 w-4 text-charcoal-400" />
-          <AlertTitle>Comparison unavailable</AlertTitle>
-          <AlertDescription className="font-poppins text-charcoal-500">
-            {comparisonState.message}
-          </AlertDescription>
-        </Alert>
-      )}
-
-      {comparisonState.kind === 'data' &&
-        comparisonState.comparison.comparisonStatus === 'no_published_version' && (
-          <Alert className="border-beige-200">
-            <AlertCircle className="h-4 w-4 text-charcoal-400" />
-            <AlertTitle>No published version yet</AlertTitle>
-            <AlertDescription className="font-poppins text-charcoal-500">
-              Publish a configuration to unlock publish-to-publish comparison.
-            </AlertDescription>
-          </Alert>
-        )}
-
-      {comparisonState.kind === 'data' &&
-        comparisonState.comparison.comparisonStatus === 'no_previous_version' &&
-        comparisonState.comparison.currentVersion && (
-          <>
-            <div className="rounded-md border border-beige-200 p-4 space-y-3">
-              <div>
-                <p className="font-medium text-charcoal">
-                  Current Published Version v{comparisonState.comparison.currentVersion.version}
-                </p>
-                <p className="text-sm text-charcoal-500 font-poppins">
-                  Published{' '}
-                  {formatDateOrFallback(comparisonState.comparison.currentVersion.publishedAt)}
-                </p>
-              </div>
-              {renderRunSummary(comparisonState.comparison.currentVersion)}
-            </div>
-
-            <Alert className="border-beige-200">
-              <AlertCircle className="h-4 w-4 text-charcoal-400" />
-              <AlertTitle>Previous version unavailable</AlertTitle>
-              <AlertDescription className="font-poppins text-charcoal-500">
-                Publish at least two versions to see metric deltas between releases.
-              </AlertDescription>
-            </Alert>
-          </>
-        )}
-
-      {comparisonState.kind === 'data' &&
-        comparisonState.comparison.comparisonStatus === 'comparable' &&
-        comparisonState.comparison.currentVersion &&
-        comparisonState.comparison.previousVersion && (
-          <>
-            <div className="grid gap-4 md:grid-cols-2">
-              <div className="rounded-md border border-beige-200 p-4 space-y-3">
-                <div>
-                  <p className="font-medium text-charcoal">
-                    Current Published Version v{comparisonState.comparison.currentVersion.version}
-                  </p>
-                  <p className="text-sm text-charcoal-500 font-poppins">
-                    Published{' '}
-                    {formatDateOrFallback(comparisonState.comparison.currentVersion.publishedAt)}
-                  </p>
-                </div>
-                {renderRunSummary(comparisonState.comparison.currentVersion)}
-              </div>
-
-              <div className="rounded-md border border-beige-200 p-4 space-y-3">
-                <div>
-                  <p className="font-medium text-charcoal">
-                    Previous Published Version v{comparisonState.comparison.previousVersion.version}
-                  </p>
-                  <p className="text-sm text-charcoal-500 font-poppins">
-                    Published{' '}
-                    {formatDateOrFallback(comparisonState.comparison.previousVersion.publishedAt)}
-                  </p>
-                </div>
-                {renderRunSummary(comparisonState.comparison.previousVersion)}
-              </div>
-            </div>
-
-            <div className="grid gap-4 md:grid-cols-2 xl:grid-cols-4">
-              {comparisonState.comparison.metricDeltas.map((delta) => (
-                <div
-                  key={delta.metric}
-                  className="rounded-md border border-beige-200 bg-beige-50 p-4 space-y-2"
-                >
-                  <p className="text-xs text-charcoal-400 font-poppins">{delta.displayName}</p>
-                  <p className="text-lg font-medium text-charcoal">
-                    {formatComparisonMetricValue(delta.metric, delta.currentValue)}
-                  </p>
-                  <p className="text-sm text-charcoal-500 font-poppins">
-                    Previous {formatComparisonMetricValue(delta.metric, delta.previousValue)}
-                  </p>
-                  {delta.driftCapable ? (
-                    <p className="text-sm font-medium text-charcoal">
-                      Delta {formatComparisonDelta(delta)}
-                    </p>
-                  ) : (
-                    <div className="space-y-1">
-                      <p className="text-sm font-medium text-charcoal">Drift unavailable</p>
-                      <p className="text-xs text-charcoal-500 font-poppins">
-                        {formatDriftCapabilityReason(delta)}
-                      </p>
-                    </div>
-                  )}
-                </div>
-              ))}
-            </div>
-          </>
-        )}
-    </div>
-  );
-}
-
-function RecalcButton({
-  lifecycle,
-  recalculateState,
-  onRecalculate,
-}: {
-  lifecycle: FundStateReadV1;
-  recalculateState: RecalculateState;
-  onRecalculate: () => void;
-}) {
-  const hasPublished = lifecycle.configState.hasPublished;
-  const status = lifecycle.calculationState.status;
-  const calculationInProgress = status === 'submitted' || status === 'calculating';
-  const disabled = !hasPublished || calculationInProgress || recalculateState.kind === 'submitting';
-
-  let helperText = 'Re-run calculations for the current published configuration.';
-  if (!hasPublished) {
-    helperText = 'Publish a configuration before recalculating results.';
-  } else if (calculationInProgress) {
-    helperText = 'Calculation is already in progress for the published configuration.';
-  } else if (recalculateState.kind === 'submitting') {
-    helperText = 'Starting recalculation for the published configuration.';
-  }
-
-  return (
-    <div className="flex flex-col items-start gap-2 md:items-end">
-      <Button
-        onClick={onRecalculate}
-        disabled={disabled}
-        variant="outline"
-        data-testid="recalculate-button"
-      >
-        <RefreshCw
-          className={cn('h-4 w-4', recalculateState.kind === 'submitting' && 'animate-spin')}
-        />
-        {recalculateState.kind === 'submitting' ? 'Starting Recalculation…' : 'Recalculate'}
-      </Button>
-      <p className="text-xs text-charcoal-400 font-poppins">{helperText}</p>
-    </div>
-  );
-}
-
-function LifecycleStatusCard({
-  lifecycle,
-  recalculateState,
-  onRecalculate,
-}: {
-  lifecycle: FundStateReadV1;
-  recalculateState: RecalculateState;
-  onRecalculate: () => void;
-}) {
-  const { configState, calculationState } = lifecycle;
-  const diagnostic = getLifecycleDiagnostic(lifecycle);
-  const availableSnapshotList =
-    calculationState.availableSnapshotTypes.length > 0
-      ? calculationState.availableSnapshotTypes.join(', ')
-      : 'None yet';
-  const expectedSnapshotList =
-    calculationState.expectedSnapshotTypes.length > 0
-      ? calculationState.expectedSnapshotTypes.join(', ')
-      : 'None';
-
-  return (
-    <div
-      className="bg-white rounded-lg border border-beige-200 p-6 space-y-4"
-      data-testid="run-diagnostics-card"
-    >
-      <div className="flex flex-col gap-4 md:flex-row md:items-start md:justify-between">
-        <div>
-          <h2 className="text-lg font-medium text-charcoal">Lifecycle Status</h2>
-          <p className="text-sm text-charcoal-500 font-poppins mt-1">
-            Server-backed publication, calculation, and run diagnostics for this fund.
-          </p>
-        </div>
-
-        <RecalcButton
-          lifecycle={lifecycle}
-          recalculateState={recalculateState}
-          onRecalculate={onRecalculate}
-        />
-      </div>
-
-      <Alert className={diagnosticAlertClasses(diagnostic.tone)}>
-        <AlertCircle className="h-4 w-4 text-charcoal-500" />
-        <AlertTitle>{diagnostic.title}</AlertTitle>
-        <AlertDescription className="font-poppins text-charcoal-600">
-          {diagnostic.description}
-        </AlertDescription>
-      </Alert>
-
-      {!configState.hasPublished && (
-        <div className="rounded-md border border-beige-200 bg-white p-4">
-          <div className="flex flex-col gap-3 md:flex-row md:items-center md:justify-between">
-            <div>
-              <p className="font-medium text-charcoal">Publish the fund configuration</p>
-              <p className="mt-1 text-sm text-charcoal-500 font-poppins">
-                Review the current setup, publish it, then return here to request lifecycle-backed
-                calculations.
-              </p>
-            </div>
-            <Button asChild>
-              <Link href={`/fund-setup?step=7&fundId=${lifecycle.fundId}`}>Review and publish</Link>
-            </Button>
-          </div>
-        </div>
-      )}
-
-      <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-        <FactTile
-          label="Published Version"
-          value={
-            configState.publishedVersion != null
-              ? `v${configState.publishedVersion}`
-              : 'Not published'
-          }
-        />
-        <FactTile
-          label="Published At"
-          value={formatDateOrFallback(configState.publishedAt, 'Not published')}
-        />
-        <FactTile
-          label="Calculation Status"
-          value={formatLifecycleStatus(calculationState.status)}
-        />
-        <FactTile
-          label="Last Calculated"
-          value={formatDateOrFallback(calculationState.lastCalculatedAt)}
-        />
-      </div>
-
-      <div className="grid grid-cols-2 md:grid-cols-3 gap-4">
-        <FactTile
-          label="Draft Version"
-          value={configState.draftVersion != null ? `v${configState.draftVersion}` : 'No draft'}
-        />
-        <FactTile
-          label="Latest Run"
-          value={calculationState.runId != null ? String(calculationState.runId) : 'Not started'}
-        />
-        <FactTile
-          label="Snapshot Coverage"
-          value={`${calculationState.availableSnapshotTypes.length}/${calculationState.expectedSnapshotTypes.length}`}
-        />
-      </div>
-
-      <div className="grid gap-3 md:grid-cols-2">
-        <div className="rounded-md border border-beige-200 bg-beige-50 p-4">
-          <p className="text-xs text-charcoal-400 font-poppins">Available Snapshots</p>
-          <p className="mt-1 text-sm text-charcoal font-medium">{availableSnapshotList}</p>
-        </div>
-        <div className="rounded-md border border-beige-200 bg-beige-50 p-4">
-          <p className="text-xs text-charcoal-400 font-poppins">Expected Snapshots</p>
-          <p className="mt-1 text-sm text-charcoal font-medium">{expectedSnapshotList}</p>
-        </div>
-      </div>
-
-      {calculationState.lastError && (
-        <Alert className="border-beige-200">
-          <AlertCircle className="h-4 w-4 text-charcoal-400" />
-          <AlertTitle>Latest calculation error</AlertTitle>
-          <AlertDescription className="font-poppins text-charcoal-500">
-            {calculationState.lastError}
-          </AlertDescription>
-        </Alert>
-      )}
-
-      {recalculateState.kind === 'error' && (
-        <Alert className="border-beige-200">
-          <AlertCircle className="h-4 w-4 text-charcoal-400" />
-          <AlertTitle>Recalculation failed</AlertTitle>
-          <AlertDescription className="font-poppins text-charcoal-500">
-            {recalculateState.message}
-          </AlertDescription>
-        </Alert>
-      )}
-    </div>
-  );
-}
-
-// ============================================================================
-// SECTION RENDERER
-// ============================================================================
-
-interface SectionRendererProps {
-  title: string;
-  // Accept the Zod-inferred union types: each section is a discriminated union
-  // of available/unavailable/pending/failed variants with different shapes
-  section: {
-    status: string;
-    reason?: string | undefined;
-    reasonCode?: string | undefined;
-    payload?: unknown | undefined;
-    legacyEvidence?: boolean | undefined;
-    [key: string]: unknown;
-  };
-  renderPayload?: (payload: unknown) => React.ReactNode;
-  evidenceLifecycle?: EvidenceHeaderLifecycle | undefined;
-  evidenceTestId?: string;
-}
-
-function SectionRenderer({
-  title,
-  section,
-  renderPayload,
-  evidenceLifecycle,
-  evidenceTestId,
-}: SectionRendererProps) {
-  const evidence = sectionEvidence(evidenceLifecycle, section);
-
-  if (section.status === 'available') {
-    return (
-      <div className="bg-white rounded-lg border border-beige-200 p-6">
-        <div className="mb-4 space-y-2">
-          <h2 className="text-lg font-medium text-charcoal">{title}</h2>
-          {evidence && <EvidenceHeader lifecycle={evidence} testId={evidenceTestId} />}
-        </div>
-        {section.legacyEvidence && (
-          <p className="text-xs text-charcoal-400 mb-2">
-            Based on previous calculation (legacy data)
-          </p>
-        )}
-        {renderPayload ? (
-          renderPayload(section.payload)
-        ) : (
-          <pre className="text-sm text-charcoal-600 whitespace-pre-wrap font-mono">
-            {JSON.stringify(section.payload, null, 2)}
-          </pre>
-        )}
-      </div>
-    );
-  }
-
-  const statusLabel =
-    section.status === 'failed'
-      ? section.reasonCode === 'INVALID_PUBLISHED_CONFIG'
-        ? 'Configuration issue'
-        : 'Calculation failed'
-      : section.status === 'pending'
-        ? 'Pending'
-        : '';
-  const copy = reasonCopyFor(section);
-
-  return (
-    <div className="bg-beige-50 rounded-lg border border-beige-200 p-6">
-      <div className="mb-2 space-y-2">
-        <h2 className="text-lg font-medium text-charcoal-400">{title}</h2>
-        {evidence && <EvidenceHeader lifecycle={evidence} testId={evidenceTestId} />}
-      </div>
-      <p className="text-sm text-charcoal-500 font-poppins">
-        {statusLabel ? `${statusLabel}: ` : ''}
-        {copy}
-      </p>
-    </div>
-  );
 }
 
 // ============================================================================
